requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.3.14
langchain-text-splitters==0.3.4
google-genai==1.11.0
chromadb==0.5.23
streamlit==1.43.2
requests==2.33.1
